requirements.txt
kafka-python
apscheduler
python-dotenv
psycopg2
